Issue #508: Move URL matching functionality to system engine
