+++ components/browser/engine-system/src/test/java/mozilla/components/browser/engine/system/matcher/TrieTest.kt
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-package mozilla.components.support.utils.matcher
+package mozilla.components.browser.engine.system.matcher
 
 import org.junit.Test
 import org.junit.runner.RunWith
